requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1
python-dotenv==1.0.0
bleach==6.1.0
flask-limiter==3.5.0
werkzeug==2.3.7
gunicorn==21.2.0
